BezierCurve.[min|max]_point()
Implementing [min|max]_point() so that .bounds() doesn't choke. Works by
tracing out the entire curve, then fitting a box to the points drawn. I
believe the bounds could be set to enclose the end points and control
points, if a too-large box is acceptable as a trade-off for simplicity.

Another possible improvement would be just enlarging the box by one
point all around, to make sure "fractional points" that are lost to
rounding wouldn't be clipped out by something down the road.