added unassigned shapes to virtual component
created a new component with ID 'UNASSIGNED_SHAPES' that contains
all shapes that are not assigned to any component in the
schematic. The inverse functionality is added to the writer with
extracting shapes from this virtual component and setting it as
basic shapes in the schematic file (without any component
reference).