clean up in GEDA writer/parser and corresponding tests
I cleaned up the code according to PEP8 and pylint in both GEDA
parser and writer. In addition, I fixed up the test for both
modules.
I also corrected an issue with naming nets and nets that have no
name assigend to them. Naming is now ignored when no segment in the
net has a name assigned to it (in GEDA).