created new command objects to reflect GEDA commands
In an effort to reduce the amount duplicated data I created new
classes corresponding to the GEDA commands. The commands provide
a list of parameters that are modelled as classes, too.

The modelling of commands as classes has the advantage, that the
commands used in GEDA can be generated automatically from from the
list of named parameters. This reduces the amount of hardcoded
GEDA commands, types and parameters.

In addition, I introducted ``GEDAParameter`` and ``GEDAStyleParameter``
to provide easier access to the parameters, their names, types and
default values. ``GEDAStyleParameter`` is derived from ``GEDAParameter``
and differs only in the way its name is represented (prefixed with
*style_*) to avoid collision of duplicate names for parameters.

This commit provide changes in the parser but does not yet cover
the writer.