implemented writer-side of new commands
I implemented the new GEDA command structure on the writer-side.
The GEDA commands use default settings for style that are used in
the generation of GEDA command within the writer. This allows for
easy integration of the stored style data.

In addition, I realised a minor error in offsetting the origin and
correct it by storing offset value in the design attributes. These
attributes will be extracted on the writer side to generate close
to equal coordinates in GEDA.