added improved path parsing and improved test coverage
I re-implemented the conversion of GEDA path commands into
OpenJSON shapes. Each path is now tracked in the parser with
a path ID that gets stored in the ``styles`` dictionary of each
shape. On the writer side all shapes get sorted by path ID and
are then processed as paths based on ID. Shapes without a path ID
are converted into individual shapes.

I also improved test coverage and documentation in both the writer
and parser.