Refactoring inside image.py
I'm aiming to have better division of responsibility/ownership for the
image writer. To this end, there is now a Worker class that does all the
drawing work. It encapsulates the design and canvas to be rendered to
(streamlining the arguments when calling the draw_ methods), while the
original Image class holds details about the output format. This will
allow eg. having one Image instance for each style of rendered output
desired, and just passing a number of designs through it, without
littering the Image object with leftover state.